refactor: extract shared saga runtime module
