+++ buildSrc/src/main/kotlin/net/firedevops/firemud/FiremudServiceConventionsPlugin.kt
@@ -19,6 +19,7 @@ class FiremudServiceConventionsPlugin : Plugin<Project> {
             dependencies.add("implementation", libs.findLibrary("spring.boot.starter").get())
             dependencies.add("implementation", libs.findLibrary("spring.boot.starter.actuator").get())
             dependencies.add("implementation", project.project(":common-library"))
+            dependencies.add("implementation", project.project(":common-saga"))
             dependencies.add("implementation", libs.findLibrary("grpc.spring.boot.starter").get())
             dependencies.add("implementation", libs.findLibrary("micrometer.core").get())
             dependencies.add("implementation", libs.findLibrary("micrometer.registry.prometheus").get())
