+++ services/common-saga/src/main/java/net/firedevops/firemud/common/config/CommonSagaAutoConfiguration.java
@@ -1,0 +1,31 @@
+package net.firedevops.firemud.common.config;
+
+import io.micrometer.core.instrument.MeterRegistry;
+import net.firedevops.firemud.common.saga.SagaRunner;
+import net.firedevops.firemud.common.saga.persistence.SagaInstanceRepository;
+import net.firedevops.firemud.common.saga.persistence.SagaStepRepository;
+import net.firedevops.firemud.metrics.SagaMetrics;
+import org.springframework.boot.autoconfigure.AutoConfiguration;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
+
+@AutoConfiguration
+public class CommonSagaAutoConfiguration {
+
+  @Bean
+  @ConditionalOnMissingBean
+  public SagaMetrics sagaMetrics(MeterRegistry registry) {
+    return new SagaMetrics(registry);
+  }
+
+  @Bean
+  @ConditionalOnBean({SagaInstanceRepository.class, SagaStepRepository.class})
+  @ConditionalOnMissingBean
+  public SagaRunner sagaRunner(
+      SagaMetrics sagaMetrics,
+      SagaInstanceRepository instanceRepository,
+      SagaStepRepository stepRepository) {
+    return new SagaRunner(sagaMetrics, instanceRepository, stepRepository);
+  }
+}
