+++ settings.gradle.kts
@@ -13,6 +13,9 @@ project(":common-platform-core").projectDir = File("services/common-platform-cor
 include("common-security")
 project(":common-security").projectDir = File("services/common-security")
 
+include("common-saga")
+project(":common-saga").projectDir = File("services/common-saga")
+
 include("account-service")
 project(":account-service").projectDir = File("services/account-service")
